Modularized online file fetching to its own module. Can take care of both http/https
