+++ lib/get-urls.js
@@ -9,7 +9,7 @@ module.exports = async (options) => {
   if (options.file === false) {
     var fileType = options.url.toLowerCase().split('.').pop()
     if (['txt', 'json', 'xml'].includes(fileType)) {
-      data = await urlsFromFileOnline(options.url, fileType, options.cacheTime)
+      data = await urlsFromFileOnline(options.url, options.cacheTime)
     } else {
       data = await urlsFromCrawler(options.url, options.cacheTime, options.debug)
     }
